+++ src/items/NodeItem.h
@@ -30,6 +30,8 @@ public:
     void setDisplayName(const QString& name);
     const QString& groupId() const;
     void setGroupId(const QString& groupId);
+    const QString& layerId() const;
+    void setLayerId(const QString& layerId);
 
     PortItem* addPort(const QString& portId, const QString& name, PortDirection direction);
     const QVector<PortItem*>& inputPorts() const;
@@ -64,5 +66,6 @@ private:
     QVector<PortItem*> m_outputPorts;
     QVector<PropertyData> m_properties;
     QString m_groupId;
+    QString m_layerId;
     QPointF m_dragStartPos;
 };